requirements.txt
streamlit
google-generativeai
numpy
pandas
scikit-learn
matplotlib
scipy
seaborn
